+++ DeviceOffboardingManager/Private/Config/Get-DefenderIntegrationEnabled.ps1
@@ -1,0 +1,7 @@
+function Get-DefenderIntegrationEnabled {
+    if (-not $script:DeviceOffboardingSettings) {
+        $script:DeviceOffboardingSettings = Get-DeviceOffboardingSettings
+    }
+
+    return [bool]$script:DeviceOffboardingSettings.DefenderIntegrationEnabled
+}